Inappropriate Paralleling of Countries and Regions
China and Taiwan cannot be paralleled directly under “One China” Principle based on the United Nations general assembly resolution 2758. We may parallel Mainland of China and Taiwan, China with other countries and regions.
If this is not merged. I will lead other Chinese to protest and boycott against this program.